feat(sui-studio): create Raw canvas
