+++ packages/sui-studio/workbench/src/app.js
@@ -4,7 +4,23 @@ import ReactDOM from 'react-dom'
 
 import './styles/index.scss'
 import Root from './components/Root'
+import Raw from './components/Raw'
 import {isFunction} from '../../src/components/demo/utilities'
+
+const queryStringToJSON = queryString => {
+  if (queryString.indexOf('?') > -1) {
+    queryString = queryString.split('?')[1]
+  }
+  var pairs = queryString.split('&')
+  var result = {}
+  pairs.forEach(function(pair) {
+    pair = pair.split('=')
+    result[pair[0]] = decodeURIComponent(pair[1] || '')
+  })
+  return result
+}
+
+const params = queryStringToJSON(window.location.href)
 
 const importAll = requireContext => requireContext.keys().map(requireContext)
 ;(async () => {
@@ -33,19 +49,37 @@ const importAll = requireContext => requireContext.keys().map(requireContext)
     DemoComponent = comp.default || comp
   } catch (e) {}
 
+  let demoStyles = ''
+  try {
+    demoStyles = require('!css-content-loader!css-loader!sass-loader!demo/demo/index.scss') // eslint-disable-line
+  } catch (e) {}
+
   const contexts = isFunction(ctxt) ? await ctxt() : ctxt
   const themes = requireContextThemesKeys.reduce((acc, path, index) => {
     acc[path.replace('./', '').replace('.scss', '')] = styles[index]
     return acc
   }, {})
 
+  const {raw} = params
   ReactDOM.render(
-    <Root
-      contexts={contexts}
-      themes={{...themes, default: defaultStyle.default}}
-      componentID={__COMPONENT_ID__}
-      demo={DemoComponent}
-    />,
+    !raw ? (
+      <Root
+        contexts={contexts}
+        themes={{...themes, default: defaultStyle.default}}
+        componentID={__COMPONENT_ID__}
+        demo={DemoComponent}
+        demoStyles={demoStyles}
+      />
+    ) : (
+      <Raw
+        contexts={contexts}
+        themes={{...themes, default: defaultStyle.default}}
+        componentID={__COMPONENT_ID__}
+        demo={DemoComponent}
+        demoStyles={demoStyles}
+        {...params}
+      />
+    ),
     document.getElementById('app')
   )
 })()
