+++ packages/sui-studio/workbench/src/components/Root/index.scss
@@ -3,6 +3,10 @@
   display: flex;
   flex-direction: column;
   overflow-y: auto;
+
+  &-center {
+  }
+
   &-top {
     box-shadow: 0 3px 5px rgba(100, 100, 100, .49);
     padding-bottom: 8px;
